requirements.txt
discord.py>=0.16.0
pyyaml>=3.12
